+++ CHANGELOG.md
@@ -25,8 +25,8 @@ more .NET types.
 
 ### 2.0.0 (not yet released)
 * BREAKING CHANGES:
-	* bootstrap_contracts return type includes target_curve attribute.
 	* max_smooth_interp return type includes spline_parameters attribute.
 * Include target_curve parameter to bootstrap_contracts function.
+* Added return_target_curve parameter to bootstrap_contracts function.
 * Add tension parameter to maximum smoothness spline.
 * Add hyberbolic_tension_spline function.